+++ tezt/lib_tezos/baker.mli
@@ -126,8 +126,10 @@ val liquidity_baking_votefile : ?path:string -> liquidity_baking_vote -> string
     is not passed. If it is [Some x] then [--liquidity-baking-toggle-vote x] is
     passed. The default value is [Some Pass].
 
-    [operations_pool], [force_apply] and [state_recorder] are passed to the baker 
-    daemon through the flag [--operations-pool], [--force_apply] and [--record-state].
+    [operations_pool], [force_apply], [state_recorder] and
+    [node_version_check_bypass] are passed to the baker daemon through the flag
+    [--operations-pool], [--force_apply], [--record-state] and
+    [--node-version-check-bypass].
 
     If [remote_mode] is specified, the baker will run in RPC-only mode.
 
@@ -154,6 +156,7 @@ val create :
   ?dal_node:Dal_node.t ->
   ?minimal_nanotez_per_gas_unit:int ->
   ?state_recorder:bool ->
+  ?node_version_check_bypass:bool ->
   Node.t ->
   Client.t ->
   t
@@ -192,6 +195,7 @@ val create_from_uris :
   ?dal_node_rpc_endpoint:Endpoint.t ->
   ?minimal_nanotez_per_gas_unit:int ->
   ?state_recorder:bool ->
+  ?node_version_check_bypass:bool ->
   base_dir:string ->
   node_data_dir:string ->
   node_rpc_endpoint:Endpoint.t ->
@@ -228,10 +232,12 @@ val create_from_uris :
     baker. This defaults to the empty list, which is a shortcut for "every known
     account".
 
-    [votefile], [liquidity_baking_toggle_vote], [force_apply], [state_recorder]
-    respectively [operations_pool] are passed to the baker daemon through the flags
-    [--votefile], [--liquidity-baking-toggle-vote], [--should-apply], [--record-state]
-    respectively [--operations-pool].
+    [votefile], [liquidity_baking_toggle_vote], [force_apply],
+    [state_recorder], [node_version_check_bypass] respectively
+    [operations_pool] are passed to the baker daemon through the flags
+    [--votefile], [--liquidity-baking-toggle-vote], [--should-apply],
+    [--record-state], [--node-version-check-bypass] respectively
+    [--operations-pool].
 
     If [remote_mode] is specified, the baker will run in RPC-only mode.
 
@@ -256,6 +262,7 @@ val init :
   ?dal_node:Dal_node.t ->
   ?minimal_nanotez_per_gas_unit:int ->
   ?state_recorder:bool ->
+  ?node_version_check_bypass:bool ->
   Node.t ->
   Client.t ->
   t Lwt.t